style: 버킷 메서드명 변경
